+++ src/main/java/com/documentation/volume/repository/FacultyRepository.java
@@ -2,7 +2,7 @@ package com.documentation.volume.repository;
 
 import com.documentation.volume.model.Faculty;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface FacultyRepository extends JpaRepository<Faculty, Integer> {
+public interface FacultyRepository extends JpaRepository<Faculty, Long> {
 
 
 }
