+++ src/main/java/com/documentation/volume/repository/VersionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VersionRepository extends JpaRepository<Version,Integer> {
+public interface VersionRepository extends JpaRepository<Version,Long> {
 
-    List<Version> findAllByDepartmentId(Integer Id);
+    List<Version> findAllByDepartmentId(Long Id);
 }
